requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.2
requests>=2.32.3
google-api-python-client>=2.146.0
google-auth>=2.34.0
slack-sdk>=3.33.1